backend/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
python-socketio==5.8.0
osmnx==1.3.0
networkx==3.1
numpy==1.24.3
pydantic==1.10.7
